Crate amq_protocol_types

AMQP types manipulation library

amq-protocol-types is a library aiming at providing an implementation/abstraction around AMQP types.

It implements the list of the different kind of types available, a value holder and serialization.deserialization facilities.

Modules

flags

Helpers to handle AMQP flags.

generation

Generation utilities for the various AMQP types.

parsing

Parsing utilities for the various AMQP types.

Structs

DecimalValue

A Decimal value composed of a scale and a value

Enums

AMQPType

Enumeration referencing all the available AMQP types

AMQPValue

Enumeration referencing the possible AMQP values depending on the types

Type Aliases

Boolean

A bool

ByteArray

An array of bytes (RabbitMQ specific)

Double

A f64

FieldArray

An array of AMQPValue

FieldTable

A Map<String, AMQPValue>

Float

A f32

LongInt

An i32

LongLongInt

An i64

LongLongUInt

A u64

LongString

A String

LongUInt

A u32

ShortInt

An i16

ShortShortInt

An i8

ShortShortUInt

A u8

ShortString

A String (deprecated)

ShortUInt

A u16

Timestamp

A timestamp (u32)

Void

No value